2 players, size changing, zooming
I added zooming in it's infancy, it still needs to be tweaked a bit before it can be truly viable, but the basic functionality is there. Also deleted weegee and instead made it so there were multiple players.... just 2 for now. One last thing is that the 60 and 40 numbers of how the size of a player is determined is now a variable so the player can now be up to 80 size in either sizes,